moved button into index.tsx
